chore: update project dirs, files and types
Reference-to: #48, #51, #49
