--- src/@types/events.ts
@@ -1,8 +1,0 @@
-/*===============================*
-          EVENT TYPES
- *===============================*
-*/
-export type ReactFormEvent = React.FormEvent<HTMLFormElement>;
-export type ReactSelectEvent = React.MouseEvent<HTMLLIElement>;
-export type ReactInputEvent = React.ChangeEvent<HTMLInputElement>;
-export type ReactMouseEvent = React.MouseEvent<HTMLButtonElement>;
