--- src/@types/index.ts
@@ -1,5 +1,0 @@
-export * from './events';
-export * from './helpers';
-export * from './models';
-export * from './project.d';
-export * from './types';
