+++ src/components/atoms/form-control.tsx
@@ -1,4 +1,4 @@
-import type { Props } from '@src/@types';
+import type { Project } from '@src/@types';
 import clsx from 'clsx';
 
 const FormControl = <E extends React.ElementType = 'div'>({
@@ -6,7 +6,7 @@ const FormControl = <E extends React.ElementType = 'div'>({
   className,
   children,
   ...rest
-}: Props<E>) => {
+}: Project.ElementProps<E>) => {
   const RenderedElement = as || 'div';
 
   return (
