+++ src/components/atoms/text.tsx
@@ -1,10 +1,10 @@
-import type { Props } from '@src/@types';
+import type { Project } from '@src/@types';
 
 const Text = <E extends React.ElementType = 'p'>({
   children,
   as,
   ...rest
-}: Props<E>) => {
+}: Project.ElementProps<E>) => {
   const RenderedElement = as || 'p';
   return <RenderedElement {...rest}>{children}</RenderedElement>;
 };
